// features/arrivals/types.ts
export interface Arribo {
    DescripcionLinea?: string;
    descripcionLinea?: string;
    DescripcionBandera?: string;
    descripcionBandera?: string;
    DescripcionCartelBandera?: string;
    descripcionCartelBandera?: string;
    DescripcionCortaBandera?: string;
    descripcionCortaBandera?: string;
    Arribo: string;
    CodigoLineaParada: string;
    DesvioHorario: string;
    EsAdaptado: string;
    IdentificadorChofer: string;
    IdentificadorCoche: string;
    Latitud: string;
    LatitudParada: string;
    Longitud: string;
    LongitudParada: string;
    UltimaFechaHoraGPS: string;
    MensajeError: string;
}